fix fitness function bug
fixed a bug which messed up the computation of the fitness function. Because the array which contained the difference between the two images was declared as uint8 it could not hold the values when squaring the distance. I changed the format of the array to int

also changed the config to be more explorative