adapt the test to the new tessellated polygon format
when I render the image using the training feedback and pygame it somehow draws only the half of the image. To still confirm that the rendering after the tessellation works correctly I adapted my rendering test to work with the new polygons. Therefore I also add a decomposition of the alarm clock image in the tessellated version to the test_data directory