fix import and remove duplicated file
fix the import in the TraningFeedback file. The import should be done by the full name starting from the module name. This works then on computers where the pyff module is installed somewhere in the path.

furthermore remove the pool file which was duplicated from the genetics subfolder. Not used in TrainingFeedback anyway.