clean the helper functions
remove all the old functions which were used for Luis' initial data format. But now we store everything in json and don't need these functions to read from this data format. 

I also removed functions which are only one-liners and not used regularly.

changed the comments in a way that I removed information obvious from the code and formatted the extremely long lines..