better mutation selection
select offspring also if it is not better, but equally good. This gives better results (faster convergence). I think it helps to explore the search space quicker. There was a good mathematical argument for it, but I forgot it. Heard it in the talk of the part time scientists on some genetic programming on the 28c3.