fix an issue in how we encode the stimulus in the Trigger
We realized that the Trigger send to us by the classifier does not help us to determine the right polygon because we always drew a random polygon from a certain non-target decomposition. The solution is to not use a random polygon of the non-target stimulus, but also the Nth one according to the ranking. The same way we do it with the target. Than the '1 out of 6 classes' coding of the classifier makes sense