rename dataFolder to data_path and make it a property
making it a property makes the initialization easier. I wanted to load all the files from disk already in the init of the feedback, but this was not possible as the path to the data is set later. I don't know how to set it via pyff in the init. 

Now the feedback base class immediately loads the data after the path is set..