Double the size of the images.
The size of the images is now double as it was before. 
A problem arises with the upper edge of the image, which displays a grayline over the pic which can not be removed in any obvious way so far. Same thing happened before and happens as well with the right edge of the pic, but these cases were solved by setting a stimulus size a pixel smalled than the actual image. Now this approach does not work.