fix target selection bug
Testing the TrainingFeedback it I don't see the target image more often than random. And then the strange thing, in the log I can see something like:

   2012-05-07 10:00:30 DEBUG: TARGET 1 selected for display.
   2012-05-07 10:00:30 DEBUG: Polygon 5 selected for display.

but the presented target has only 3 polygons after pruning the ones
with low error.

the problem was that the polygon pool was loaded again without pruning